backend/services/audio_service/requirements.txt
deepgram-sdk==2.12.0
uvicorn
fastapi
yt-dlp
